image-generate/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
cloudinary==1.39.1
colorama==0.4.6
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_resources==6.1.3
limits==3.10.0
motor==3.3.2
openai==1.13.3
packaging==23.2
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
wrapt==1.16.0
